+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fixed issues generating CSS for dashed strokes in shapes and borders
 - `contextmenu` builder returns `ContextMenu` instance instead of `EventTarget` (https://github.com/edvin/tornadofx/issues/702)
 - `togglebutton` and `radiobutton` builders moved from `Node` to `EventTarget` (https://github.com/edvin/tornadofx/issues/716)
+- Workspace now undocks closed tabs correctly (https://github.com/edvin/tornadofx/issues/718)
 
 ### Changes